% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_fcm.R
\name{apply_fcm_clustering}
\alias{apply_fcm_clustering}
\title{Apply Fuzzy C-Means Clustering}
\usage{
apply_fcm_clustering(data, centers = 4, m = 2)
}
\arguments{
\item{data}{data.frame with numeric columns wind_speed and power.}

\item{centers}{integer number of clusters.}

\item{m}{fuzzifier parameter (default 2).}
}
\value{
tibble with cluster assignments.
}
\description{
Apply Fuzzy C-Means Clustering
}
