% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metrics.R
\name{evaluate_metrics}
\alias{evaluate_metrics}
\title{Compute Evaluation Metrics}
\usage{
evaluate_metrics(data)
}
\arguments{
\item{data}{tibble with columns power, pred.}
}
\value{
list of RMSE, MAE, MAPE, R2, CA.
}
\description{
Compute Evaluation Metrics
}
