% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_results}
\alias{plot_results}
\title{Plot Cleaned Power Curve}
\usage{
plot_results(data)
}
\arguments{
\item{data}{tibble with columns wind_speed, power, pred.}
}
\value{
ggplot object.
}
\description{
Plot Cleaned Power Curve
}
