% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windCleanHybrid-package.R
\docType{package}
\name{windCleanHybrid-package}
\alias{windCleanHybrid}
\alias{windCleanHybrid-package}
\title{windCleanHybrid package}
\description{
Utilities for cleaning and modelling wind-turbine SCADA data with a hybrid
pipeline based on pre-cleaning, fuzzy c-means clustering, Mahalanobis
distance, and ANN refinement.
}
\author{
\strong{Maintainer}: Diogo Ribeiro \email{diogo_dj@hotmail.com}

}
\keyword{internal}
